Extracted test_suite_commtrack
